Wrap bfs in a Ruby command-line tool
Replaces the shell "wrapper.sh" script with a Ruby command-line tool
interface using the thor library. The underlying functionality is unchanged,
but the code has been reformatted to make it callable from Ruby without having
to shell out. See the updated README for details on how to invoke the app.